[export] Adjust latent_see to fill based on time since last flush
The ring buffer was inconsistently removing items making it hard to reason about the state in general.

Additionally adjust API such that we'll guarantee concurrent one flush and fail to collect others. This ultimately makes it a lot easier to limit the impact of running this system.

----
DO NOT SUBMIT. This PR is for testing purposes only. [cl/688680363](http://cl/688680363)

PiperOrigin-RevId: 688680363